// src/types/home.ts
export interface Course {
  title: string;
  description: string;
  imageSrc: string;
  imageAlt: string;
  link: string;
}

export interface CarouselItem {
  title: string;
  description: string;
  imageSrc: string;
  imageAlt: string;
}

export interface ImpactStat {
  value: string;
  title: string;
  description: string;
}